+++ srv/src/main/java/com/sap/cap/ai2code/service/interfaces/BotExecution.java
@@ -1,3 +1,4 @@
+package com.sap.cap.ai2code.service.interfaces;
 // package com.sap.cap.ai2code.service;
 
 // public interface BotExecution {
