+++ srv/src/main/java/com/sap/cap/ai2code/service/interfaces/ContextService.java
@@ -1,3 +1,4 @@
+package com.sap.cap.ai2code.service.interfaces;
 // package com.sap.cap.ai2code.service;
 
 // import java.util.List;
