+++ srv/src/main/java/com/sap/cap/ai2code/service/interfaces/RAGExtractor.java
@@ -1,3 +1,4 @@
+package com.sap.cap.ai2code.service.interfaces;
 // package com.sap.cap.ai2code.service;
 
 // public interface RAGExtractor {
